back-end/requirements.txt
# Core - FastAPI
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6  # Form data support
python-dotenv>=1.0.0,<2.0.0

# Security and Performance
slowapi>=0.1.9,<1.0.0    # Rate limiting for FastAPI
pyjwt>=2.8.0,<3.0.0      # JWT tokens

# DB - Multi-database support
mysql-connector-python>=8.1.0,<9.0.0
psycopg2-binary>=2.9.0,<3.0.0
pyodbc>=4.0.39,<5.0.0
oracledb>=2.0.0,<3.0.0
# SQLite is built into Python, no additional package needed

# Firebase Admin SDK
firebase-admin>=6.2.0,<7.0.0

# Firestore (if used directly)
google-cloud-firestore>=2.16.0,<3.0.0

# Auth
google-auth>=2.29.0,<3.0.0

# HTTP client (used by Firebase / GCP libs)
requests>=2.28.0,<3.0.0

# Redis for sessions and caching
redis>=5.0.0,<6.0.0

# LLM — LangGraph + LangChain ecosystem
langgraph>=1.1.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-redis>=0.1.0

pydantic>=2.0.0,<3.0.0
